+++ src/main/kotlin/site/iplease/iadserver/domain/demand/data/type/DemandPolicyGroupType.kt
@@ -1,5 +1,5 @@
 package site.iplease.iadserver.domain.demand.data.type
 
 enum class DemandPolicyGroupType {
-    DEMAND_CANCEL
+    DEMAND_CANCEL, DEMAND_CREATE
 }
