--- src/main/kotlin/site/iplease/iadserver/domain/demand/util/DemandDataValidator.kt
@@ -1,8 +1,0 @@
-package site.iplease.iadserver.domain.demand.util
-
-import reactor.core.publisher.Mono
-import site.iplease.iadserver.domain.demand.data.dto.DemandDto
-
-interface DemandDataValidator {
-    fun validate(dto: DemandDto): Mono<DemandDto>
-}
