+++ src/main/kotlin/site/iplease/iadserver/domain/demand/util/DemandPolicyValidator.kt
@@ -1,8 +1,9 @@
 package site.iplease.iadserver.domain.demand.util
 
 import reactor.core.publisher.Mono
+import site.iplease.iadserver.domain.demand.data.dto.DemandDto
 import site.iplease.iadserver.domain.demand.data.type.DemandPolicyGroupType
 
 interface DemandPolicyValidator {
-    fun validate(demandId: Long, accountId: Long, policy: DemandPolicyGroupType): Mono<Unit>
+    fun validate(demand: DemandDto, policy: DemandPolicyGroupType): Mono<DemandDto>
 }
